highway implementation

+++ include/Simulator.h
@@ -14,7 +14,7 @@ class Simulator
 {
     public:
         Simulator(const Highway& highway);
-        std::vector<Interval> generateVelProfile(int distanceKm, double currentDepartureTime);
+        std::vector<Interval> generateVelProfile(double distanceKm, double currentDepartureTime);
         void generateRuns(); // writes on Runs.txt file
         void generatePassages(); // writes on Passages.txt file
